Add tile dimensions into snapshot and use it for setting tile rect
Current tile rendering implementation in hwui sets the current
clip rect as tile rect during restore. This will limit the rendering
to the restored clip rect. Later in case if the app tries to render
outside the tile rect by setting a new clip rect, then all the pixels,
which are falling outside the tile rect are clipped off, this may result
into UI artifacats. This change adds the support for preserving the tile
clip as part of the snapshot and the same tile clip is used while
setting the tile rect. This way it is taken care that the correct tile
dimensions are preserved during save and retrieved during restore.

CRs-Fixed: 499767
Change-Id: Ia66b6dc8e7be5857949751a81e9f702c2d1c5a57